Fixes
Some small change in the semantic checker. Also fixed a previous issue with `asm`.

+++ src/vp_asm.h
@@ -71,7 +71,7 @@ typedef struct AsmOperand
     AsmOperandType type;
     union
     {
-        uint8_t reg;
+        uint64_t reg;
         int64_t imm;
         Str* ident;
         struct
